requirements.txt
Django
djangorestframework
gunicorn
whitenoise
psycopg2-binary
dj-database-url
python-dotenv
Pillow
